requirements.txt
streamlit>=1.33.0
anthropic>=0.21.3
python-dotenv>=1.0.1
rich>=13.7.1
